add categories file and move categories code
